small reorg of pathEndRuntime to its own file
