+++ relayer/processor/path_processor.go
@@ -47,75 +47,11 @@ func (p PathProcessors) IsRelayedChannel(k ChannelKey, chainID string) bool {
 	return false
 }
 
-// pathEndRuntime is used at runtime for each chain involved in the path.
-// It holds a channel for incoming messages from the ChainProcessors, which are
-// processed during Run(ctx).
-type pathEndRuntime struct {
-	info PathEnd
-
-	chainProvider provider.ChainProvider
-
-	// cached data
-	latestBlock          provider.LatestBlock
-	messageCache         IBCMessagesCache
-	clientState          provider.ClientState
-	clientTrustedState   provider.ClientTrustedState
-	connectionStateCache ConnectionStateCache
-	channelStateCache    ChannelStateCache
-	latestHeader         provider.IBCHeader
-	ibcHeaderCache       IBCHeaderCache
-
-	// New messages and other data arriving from the handleNewMessagesForPathEnd method.
-	incomingCacheData chan ChainProcessorCacheData
-
-	// Cache in progress sends for the different IBC message types
-	// to avoid retrying a message immediately after it is sent.
-	packetSendCache     packetSendCache
-	connectionSendCache connectionSendCache
-	channelSendCache    channelSendCache
-
-	// inSync indicates whether queries are in sync with latest height of the chain.
-	inSync bool
-}
-
-func (pathEnd *pathEndRuntime) MergeCacheData(d ChainProcessorCacheData) {
-	pathEnd.inSync = d.InSync
-	pathEnd.latestBlock = d.LatestBlock
-	pathEnd.latestHeader = d.LatestHeader
-	pathEnd.clientState = d.ClientState
-
-	// TODO messageCache Merge will likely need to move into the PathProcessor, and make sure passes client, connection, or channel filter
-	pathEnd.messageCache.Merge(d.IBCMessagesCache) // Merge incoming packet IBC messages into the backlog
-
-	// TODO connectionStateCache Merge will likely need to move into the PathProcessor, and make sure passes connection filter
-	pathEnd.connectionStateCache.Merge(d.ConnectionStateCache) // Update latest connection open state for chain
-
-	// TODO channelStateCache Merge will likely need to move into the PathProcessor, and make sure passes connection filter
-	pathEnd.channelStateCache.Merge(d.ChannelStateCache) // Update latest channel open state for chain
-
-	pathEnd.ibcHeaderCache.Merge(d.IBCHeaderCache)  // Update latest IBC header state
-	pathEnd.ibcHeaderCache.Prune(ibcHeadersToCache) // Only keep most recent IBC headers
-}
-
 func NewPathProcessor(log *zap.Logger, pathEnd1 PathEnd, pathEnd2 PathEnd) *PathProcessor {
 	return &PathProcessor{
-		log: log,
-		pathEnd1: &pathEndRuntime{
-			info:                 pathEnd1,
-			incomingCacheData:    make(chan ChainProcessorCacheData, 100),
-			connectionStateCache: make(ConnectionStateCache),
-			channelStateCache:    make(ChannelStateCache),
-			messageCache:         NewIBCMessagesCache(),
-			ibcHeaderCache:       make(IBCHeaderCache),
-		},
-		pathEnd2: &pathEndRuntime{
-			info:                 pathEnd2,
-			incomingCacheData:    make(chan ChainProcessorCacheData, 100),
-			connectionStateCache: make(ConnectionStateCache),
-			channelStateCache:    make(ChannelStateCache),
-			messageCache:         NewIBCMessagesCache(),
-			ibcHeaderCache:       make(IBCHeaderCache),
-		},
+		log:          log,
+		pathEnd1:     newPathEndRuntime(pathEnd1),
+		pathEnd2:     newPathEndRuntime(pathEnd2),
 		retryProcess: make(chan struct{}, 8),
 	}
 }
